ip /ip range impr (constructing ip range by "n1.n2.*" added)

+++ trunk/core/src/org/punksearch/ip/IpIterator.java
@@ -33,7 +33,7 @@ public class IpIterator implements Iterator<Ip>
 
 		Ip result = nextIp;
 
-		nextIp = new Ip(nextIp.toLong().longValue() + 1);
+		nextIp = new Ip(nextIp.toLong() + 1);
 		if (nextIp.toLong() > currentIpRange.getFinishIp().toLong())
 		{
 			if (rangesIterator.hasNext())
